Add one command to create moderator group and add users to the group
